+++ server/testserver/main.go
@@ -282,7 +282,7 @@ func wssHandler(c context.Context, ctx *app.RequestContext) {
 				// slog.ErrorContext(c, "read err:", "error", err)
 				break
 			}
-			// slog.Info("recv", "msg", string(msg))
+			//slog.Info("recv", "msg", string(msg))
 
 			err = conn.WriteMessage(websocket.TextMessage, orderResp)
 			if err != nil {